requirements.txt
websockets>=12.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
duckdb>=0.8.0
pytest>=7.0.0
polygon-api-client>=1.12.0
python-dotenv>=1.0.0
scipy>=1.10.0
fastparquet>=2023.8.0
matplotlib>=3.7.0
plotly>=5.15.0
zstandard>=0.21.0